Lexer: remove the Keyword ahead-of-time hashing
This was meant to be an optimization but for the small set of keywords Lua uses it ends up being slower than just using `mem.eql`. I'm not super confident about this. Will need to revisit once we start benchmarking with more valid Lua instead of the corpus from fuzzing the lexer (which presumably doesn't have as many valid non-keyword identifiers in its corpus as human-written source files would normally)