Lexer: Make Lexer comptime configurable
Lexer is now a function that takes a comptime config struct and returns a type. This allows us to test different configurations of the Lexer at the same time (previously, LUA_51_COMPAT_CHECK_NEXT_BUG tests would just get skipped if it was set to false) and it allows users to more easily use different lexer settings.