Add string parsing + fuzzed string test set
I followed mostly the same procedure outlined here: https://www.ryanliptak.com/blog/fuzzing-as-test-case-generator/
but used a combination of Zua and fuzzing-lua to ultimately create the sets of inputs/outputs.

- First, a giant corpus (16k+) of fuzzed string literals was created by iterating through all the fuzzed lexer inputs and outputting the source of every <string> token to a separate file (see test/fuzz_strings_gen.zig). This step would be difficult to do with Lua's API because strings are parsed as they are lexed, meaning any relationship to the original source is lost once the token is parsed.
- Then, I used libFuzzer and fuzzing-lua to minimize the string corpus (via the -merge=1 flag).
- Then, I used Lua to generate corresponding output files containing the parsed version of each input string (this code will be committed to fuzzing-lua once I clean it up).

Kind of convoluted, but it ended up working well--there were a lot of bugs in my initial string parsing implementation that the fuzzed set allowed me to find.