Undo comptime configuration of the lexer for simplicity's sake
Having things like `check_next_bug_compat` be comptime-configurable would allow for better performance (eliminating dead branches at compile time), but it would also necessitate turning Lexer into an interface in order to sanely pass the lexer around when parsing, etc. For now, I'm just going with runtime configuration of the lexer since that'll be much easier to work with, but will try to revisit this in the future.